requirements.txt
aiofiles==0.4.0
aiohttp==3.5.4
appnope==0.1.0
async-timeout==3.0.1
attrs==18.2.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
Bottleneck==1.2.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dataclasses==0.6
decorator==4.3.2
defusedxml==0.5.0
dill==0.2.9
entrypoints==0.3
fastai==1.0.46
fastprogress==0.1.19
google-images-download==2.5.0
h11==0.8.1
httptools==0.0.13
idna==2.8
idna-ssl==1.1.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
msgpack==0.5.6
msgpack-numpy==0.4.3.2
multidict==4.5.2
murmurhash==1.0.1
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.4
numexpr==2.6.9
numpy==1.16.1
nvidia-ml-py3==7.352.0
opencv-python==4.0.0.21
packaging==19.0
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
plac==0.9.6
preshed==2.0.1
prometheus-client==0.5.0
prompt-toolkit==2.0.8
ptyprocess==0.6.0
Pygments==2.3.1
pynvx==1.0.0
pyparsing==2.3.1
python-dateutil==2.8.0
python-multipart==0.0.5
pytz==2018.9
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
regex==2018.1.10
requests==2.21.0
scipy==1.2.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.7.3
spacy==2.0.18
starlette==0.11.3
terminado==0.8.1
testpath==0.4.2
thinc==6.12.1
toolz==0.9.0
torch==1.0.1.post2
torchvision==0.2.1
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
typing==3.6.6
typing-extensions==3.7.2
ujson==1.35
urllib3==1.24.1
uvicorn==0.4.6
uvloop==0.12.1
wcwidth==0.1.7
webencodings==0.5.1
websockets==7.0
widgetsnbextension==3.4.2
wrapt==1.10.11
yarl==1.3.0
